% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Account}
\alias{Account}
\title{DCM/DFA Reporting And Trafficking API Objects 
Manages your DoubleClick Campaign Manager ad campaigns and reports.}
\usage{
Account(accountPermissionIds = NULL, accountProfile = NULL, active = NULL,
  activeAdsLimitTier = NULL, activeViewOptOut = NULL,
  availablePermissionIds = NULL, countryId = NULL, currencyId = NULL,
  defaultCreativeSizeId = NULL, description = NULL, id = NULL,
  locale = NULL, maximumImageSize = NULL, name = NULL,
  nielsenOcrEnabled = NULL, reportsConfiguration = NULL,
  shareReportsWithTwitter = NULL, teaserSizeLimit = NULL)
}
\arguments{
\item{accountPermissionIds}{Account permissions assigned to this account}

\item{accountProfile}{Profile for this account}

\item{active}{Whether this account is active}

\item{activeAdsLimitTier}{Maximum number of active ads allowed for this account}

\item{activeViewOptOut}{Whether to serve creatives with Active View tags}

\item{availablePermissionIds}{User role permissions available to the user roles of this account}

\item{countryId}{ID of the country associated with this account}

\item{currencyId}{ID of currency associated with this account}

\item{defaultCreativeSizeId}{Default placement dimensions for this account}

\item{description}{Description of this account}

\item{id}{ID of this account}

\item{locale}{Locale of this account}

\item{maximumImageSize}{Maximum image size allowed for this account}

\item{name}{Name of this account}

\item{nielsenOcrEnabled}{Whether campaigns created in this account will be enabled for Nielsen OCR reach ratings by default}

\item{reportsConfiguration}{Reporting configuration of this account}

\item{shareReportsWithTwitter}{Share Path to Conversion reports with Twitter}

\item{teaserSizeLimit}{File size limit in kilobytes of Rich Media teaser creatives}
}
\value{
Account object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:46:46
filename: /Users/mark/dev/R/autoGoogleAPI/googledfareportingv27.auto/R/dfareporting_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Account Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a DCM account.
}
\seealso{
Other Account functions: \code{\link{accounts.patch}},
  \code{\link{accounts.update}}
}
