% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{AccountActiveAdSummary}
\alias{AccountActiveAdSummary}
\title{AccountActiveAdSummary Object}
\usage{
AccountActiveAdSummary(accountId = NULL, activeAds = NULL,
  activeAdsLimitTier = NULL, availableAds = NULL)
}
\arguments{
\item{accountId}{ID of the account}

\item{activeAds}{Ads that have been activated for the account}

\item{activeAdsLimitTier}{Maximum number of active ads allowed for the account}

\item{availableAds}{Ads that can be activated for the account}
}
\value{
AccountActiveAdSummary object
}
\description{
AccountActiveAdSummary Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Gets a summary of active ads in an account.
}
