% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{AccountPermissionGroup}
\alias{AccountPermissionGroup}
\title{AccountPermissionGroup Object}
\usage{
AccountPermissionGroup(id = NULL, name = NULL)
}
\arguments{
\item{id}{ID of this account permission group}

\item{name}{Name of this account permission group}
}
\value{
AccountPermissionGroup object
}
\description{
AccountPermissionGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
AccountPermissionGroups contains a mapping of permission group IDs to names. A permission group is a grouping of account permissions.
}
