% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{AccountPermissionsListResponse}
\alias{AccountPermissionsListResponse}
\title{AccountPermissionsListResponse Object}
\usage{
AccountPermissionsListResponse(accountPermissions = NULL)
}
\arguments{
\item{accountPermissions}{Account permission collection}
}
\value{
AccountPermissionsListResponse object
}
\description{
AccountPermissionsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Account Permission List Response
}
