% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{AccountsListResponse}
\alias{AccountsListResponse}
\title{AccountsListResponse Object}
\usage{
AccountsListResponse(accounts = NULL, nextPageToken = NULL)
}
\arguments{
\item{accounts}{Account collection}

\item{nextPageToken}{Pagination token to be used for the next list operation}
}
\value{
AccountsListResponse object
}
\description{
AccountsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Account List Response
}
