% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{AccountUserProfile}
\alias{AccountUserProfile}
\title{AccountUserProfile Object}
\usage{
AccountUserProfile(accountId = NULL, active = NULL,
  advertiserFilter = NULL, campaignFilter = NULL, comments = NULL,
  email = NULL, id = NULL, locale = NULL, name = NULL,
  siteFilter = NULL, subaccountId = NULL, traffickerType = NULL,
  userAccessType = NULL, userRoleFilter = NULL, userRoleId = NULL)
}
\arguments{
\item{accountId}{Account ID of the user profile}

\item{active}{Whether this user profile is active}

\item{advertiserFilter}{Filter that describes which advertisers are visible to the user profile}

\item{campaignFilter}{Filter that describes which campaigns are visible to the user profile}

\item{comments}{Comments for this user profile}

\item{email}{Email of the user profile}

\item{id}{ID of the user profile}

\item{locale}{Locale of the user profile}

\item{name}{Name of the user profile}

\item{siteFilter}{Filter that describes which sites are visible to the user profile}

\item{subaccountId}{Subaccount ID of the user profile}

\item{traffickerType}{Trafficker type of this user profile}

\item{userAccessType}{User type of the user profile}

\item{userRoleFilter}{Filter that describes which user roles are visible to the user profile}

\item{userRoleId}{User role ID of the user profile}
}
\value{
AccountUserProfile object
}
\description{
AccountUserProfile Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
AccountUserProfiles contains properties of a DCM user profile. This resource is specifically for managing user profiles, whereas UserProfiles is for accessing the API.
}
\seealso{
Other AccountUserProfile functions: \code{\link{accountUserProfiles.insert}},
  \code{\link{accountUserProfiles.patch}},
  \code{\link{accountUserProfiles.update}}
}
