% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{accountUserProfiles.patch}
\alias{accountUserProfiles.patch}
\title{Updates an existing account user profile. This method supports patch semantics.}
\usage{
accountUserProfiles.patch(AccountUserProfile, profileId, id)
}
\arguments{
\item{AccountUserProfile}{The \link{AccountUserProfile} object to pass to this method}

\item{profileId}{User profile ID associated with this request}

\item{id}{User profile ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}

Other AccountUserProfile functions: \code{\link{AccountUserProfile}},
  \code{\link{accountUserProfiles.insert}},
  \code{\link{accountUserProfiles.update}}
}
