% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{AccountUserProfilesListResponse}
\alias{AccountUserProfilesListResponse}
\title{AccountUserProfilesListResponse Object}
\usage{
AccountUserProfilesListResponse(accountUserProfiles = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{accountUserProfiles}{Account user profile collection}

\item{nextPageToken}{Pagination token to be used for the next list operation}
}
\value{
AccountUserProfilesListResponse object
}
\description{
AccountUserProfilesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Account User Profile List Response
}
