% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{AdSlot}
\alias{AdSlot}
\title{AdSlot Object}
\usage{
AdSlot(comment = NULL, compatibility = NULL, height = NULL,
  linkedPlacementId = NULL, name = NULL, paymentSourceType = NULL,
  primary = NULL, width = NULL)
}
\arguments{
\item{comment}{Comment for this ad slot}

\item{compatibility}{Ad slot compatibility}

\item{height}{Height of this ad slot}

\item{linkedPlacementId}{ID of the placement from an external platform that is linked to this ad slot}

\item{name}{Name of this ad slot}

\item{paymentSourceType}{Payment source type of this ad slot}

\item{primary}{Primary ad slot of a roadblock inventory item}

\item{width}{Width of this ad slot}
}
\value{
AdSlot object
}
\description{
AdSlot Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Ad Slot
}
