% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Advertiser}
\alias{Advertiser}
\title{Advertiser Object}
\usage{
Advertiser(accountId = NULL, advertiserGroupId = NULL,
  clickThroughUrlSuffix = NULL, defaultClickThroughEventTagId = NULL,
  defaultEmail = NULL, floodlightConfigurationId = NULL,
  floodlightConfigurationIdDimensionValue = NULL, id = NULL,
  idDimensionValue = NULL, name = NULL,
  originalFloodlightConfigurationId = NULL, status = NULL,
  subaccountId = NULL, suspended = NULL)
}
\arguments{
\item{accountId}{Account ID of this advertiser}

\item{advertiserGroupId}{ID of the advertiser group this advertiser belongs to}

\item{clickThroughUrlSuffix}{Suffix added to click-through URL of ad creative associations under this advertiser}

\item{defaultClickThroughEventTagId}{ID of the click-through event tag to apply by default to the landing pages of this advertiser's campaigns}

\item{defaultEmail}{Default email address used in sender field for tag emails}

\item{floodlightConfigurationId}{Floodlight configuration ID of this advertiser}

\item{floodlightConfigurationIdDimensionValue}{Dimension value for the ID of the floodlight configuration}

\item{id}{ID of this advertiser}

\item{idDimensionValue}{Dimension value for the ID of this advertiser}

\item{name}{Name of this advertiser}

\item{originalFloodlightConfigurationId}{Original floodlight configuration before any sharing occurred}

\item{status}{Status of this advertiser}

\item{subaccountId}{Subaccount ID of this advertiser}

\item{suspended}{Suspension status of this advertiser}
}
\value{
Advertiser object
}
\description{
Advertiser Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a DCM advertiser.
}
\seealso{
Other Advertiser functions: \code{\link{advertisers.insert}},
  \code{\link{advertisers.patch}},
  \code{\link{advertisers.update}}
}
