% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{AdvertiserGroup}
\alias{AdvertiserGroup}
\title{AdvertiserGroup Object}
\usage{
AdvertiserGroup(accountId = NULL, id = NULL, name = NULL)
}
\arguments{
\item{accountId}{Account ID of this advertiser group}

\item{id}{ID of this advertiser group}

\item{name}{Name of this advertiser group}
}
\value{
AdvertiserGroup object
}
\description{
AdvertiserGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Groups advertisers together so that reports can be generated for the entire group at once.
}
\seealso{
Other AdvertiserGroup functions: \code{\link{advertiserGroups.insert}},
  \code{\link{advertiserGroups.patch}},
  \code{\link{advertiserGroups.update}}
}
