% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{advertisers.list}
\alias{advertisers.list}
\title{Retrieves a list of advertisers, possibly filtered. This method supports paging.}
\usage{
advertisers.list(profileId, advertiserGroupIds = NULL,
  floodlightConfigurationIds = NULL, ids = NULL,
  includeAdvertisersWithoutGroupsOnly = NULL, maxResults = NULL,
  onlyParent = NULL, pageToken = NULL, searchString = NULL,
  sortField = NULL, sortOrder = NULL, status = NULL,
  subaccountId = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{advertiserGroupIds}{Select only advertisers with these advertiser group IDs}

\item{floodlightConfigurationIds}{Select only advertisers with these floodlight configuration IDs}

\item{ids}{Select only advertisers with these IDs}

\item{includeAdvertisersWithoutGroupsOnly}{Select only advertisers which do not belong to any advertiser group}

\item{maxResults}{Maximum number of results to return}

\item{onlyParent}{Select only advertisers which use another advertiser's floodlight configuration}

\item{pageToken}{Value of the nextPageToken from the previous result page}

\item{searchString}{Allows searching for objects by name or ID}

\item{sortField}{Field by which to sort the list}

\item{sortOrder}{Order of sorted results, default is ASCENDING}

\item{status}{Select only advertisers with the specified status}

\item{subaccountId}{Select only advertisers with these subaccount IDs}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}
