% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{AudienceSegmentGroup}
\alias{AudienceSegmentGroup}
\title{AudienceSegmentGroup Object}
\usage{
AudienceSegmentGroup(audienceSegments = NULL, id = NULL, name = NULL)
}
\arguments{
\item{audienceSegments}{Audience segments assigned to this group}

\item{id}{ID of this audience segment group}

\item{name}{Name of this audience segment group}
}
\value{
AudienceSegmentGroup object
}
\description{
AudienceSegmentGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Audience Segment Group.
}
