% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Browser}
\alias{Browser}
\title{Browser Object}
\usage{
Browser(browserVersionId = NULL, dartId = NULL, majorVersion = NULL,
  minorVersion = NULL, name = NULL)
}
\arguments{
\item{browserVersionId}{ID referring to this grouping of browser and version numbers}

\item{dartId}{DART ID of this browser}

\item{majorVersion}{Major version number (leftmost number) of this browser}

\item{minorVersion}{Minor version number (number after first dot on left) of this browser}

\item{name}{Name of this browser}
}
\value{
Browser object
}
\description{
Browser Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about a browser that can be targeted by ads.
}
