% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Campaign}
\alias{Campaign}
\title{Campaign Object}
\usage{
Campaign(accountId = NULL,
  additionalCreativeOptimizationConfigurations = NULL,
  advertiserGroupId = NULL, advertiserId = NULL,
  advertiserIdDimensionValue = NULL, archived = NULL,
  audienceSegmentGroups = NULL, billingInvoiceCode = NULL,
  clickThroughUrlSuffixProperties = NULL, comment = NULL,
  createInfo = NULL, creativeGroupIds = NULL,
  creativeOptimizationConfiguration = NULL,
  defaultClickThroughEventTagProperties = NULL, endDate = NULL,
  eventTagOverrides = NULL, externalId = NULL, id = NULL,
  idDimensionValue = NULL, lastModifiedInfo = NULL,
  lookbackConfiguration = NULL, name = NULL, nielsenOcrEnabled = NULL,
  startDate = NULL, subaccountId = NULL, traffickerEmails = NULL)
}
\arguments{
\item{accountId}{Account ID of this campaign}

\item{additionalCreativeOptimizationConfigurations}{Additional creative optimization configurations for the campaign}

\item{advertiserGroupId}{Advertiser group ID of the associated advertiser}

\item{advertiserId}{Advertiser ID of this campaign}

\item{advertiserIdDimensionValue}{Dimension value for the advertiser ID of this campaign}

\item{archived}{Whether this campaign has been archived}

\item{audienceSegmentGroups}{Audience segment groups assigned to this campaign}

\item{billingInvoiceCode}{Billing invoice code included in the DCM client billing invoices associated with the campaign}

\item{clickThroughUrlSuffixProperties}{Click-through URL suffix override properties for this campaign}

\item{comment}{Arbitrary comments about this campaign}

\item{createInfo}{Information about the creation of this campaign}

\item{creativeGroupIds}{List of creative group IDs that are assigned to the campaign}

\item{creativeOptimizationConfiguration}{Creative optimization configuration for the campaign}

\item{defaultClickThroughEventTagProperties}{Click-through event tag ID override properties for this campaign}

\item{endDate}{Date on which the campaign will stop running}

\item{eventTagOverrides}{Overrides that can be used to activate or deactivate advertiser event tags}

\item{externalId}{External ID for this campaign}

\item{id}{ID of this campaign}

\item{idDimensionValue}{Dimension value for the ID of this campaign}

\item{lastModifiedInfo}{Information about the most recent modification of this campaign}

\item{lookbackConfiguration}{Lookback window settings for the campaign}

\item{name}{Name of this campaign}

\item{nielsenOcrEnabled}{Whether Nielsen reports are enabled for this campaign}

\item{startDate}{Date on which the campaign starts running}

\item{subaccountId}{Subaccount ID of this campaign}

\item{traffickerEmails}{Campaign trafficker contact emails}
}
\value{
Campaign object
}
\description{
Campaign Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a DCM campaign.
}
\seealso{
Other Campaign functions: \code{\link{campaigns.insert}},
  \code{\link{campaigns.patch}},
  \code{\link{campaigns.update}}
}
