% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CampaignCreativeAssociation}
\alias{CampaignCreativeAssociation}
\title{CampaignCreativeAssociation Object}
\usage{
CampaignCreativeAssociation(creativeId = NULL)
}
\arguments{
\item{creativeId}{ID of the creative associated with the campaign}
}
\value{
CampaignCreativeAssociation object
}
\description{
CampaignCreativeAssociation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Identifies a creative which has been associated with a given campaign.
}
\seealso{
Other CampaignCreativeAssociation functions: \code{\link{campaignCreativeAssociations.insert}}
}
