% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CampaignCreativeAssociationsListResponse}
\alias{CampaignCreativeAssociationsListResponse}
\title{CampaignCreativeAssociationsListResponse Object}
\usage{
CampaignCreativeAssociationsListResponse(campaignCreativeAssociations = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{campaignCreativeAssociations}{Campaign creative association collection}

\item{nextPageToken}{Pagination token to be used for the next list operation}
}
\value{
CampaignCreativeAssociationsListResponse object
}
\description{
CampaignCreativeAssociationsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Campaign Creative Association List Response
}
