% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{campaigns.list}
\alias{campaigns.list}
\title{Retrieves a list of campaigns, possibly filtered. This method supports paging.}
\usage{
campaigns.list(profileId, advertiserGroupIds = NULL, advertiserIds = NULL,
  archived = NULL, atLeastOneOptimizationActivity = NULL,
  excludedIds = NULL, ids = NULL, maxResults = NULL,
  overriddenEventTagId = NULL, pageToken = NULL, searchString = NULL,
  sortField = NULL, sortOrder = NULL, subaccountId = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{advertiserGroupIds}{Select only campaigns whose advertisers belong to these advertiser groups}

\item{advertiserIds}{Select only campaigns that belong to these advertisers}

\item{archived}{Select only archived campaigns}

\item{atLeastOneOptimizationActivity}{Select only campaigns that have at least one optimization activity}

\item{excludedIds}{Exclude campaigns with these IDs}

\item{ids}{Select only campaigns with these IDs}

\item{maxResults}{Maximum number of results to return}

\item{overriddenEventTagId}{Select only campaigns that have overridden this event tag ID}

\item{pageToken}{Value of the nextPageToken from the previous result page}

\item{searchString}{Allows searching for campaigns by name or ID}

\item{sortField}{Field by which to sort the list}

\item{sortOrder}{Order of sorted results, default is ASCENDING}

\item{subaccountId}{Select only campaigns that belong to this subaccount}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}
