% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{cities.list}
\alias{cities.list}
\title{Retrieves a list of cities, possibly filtered.}
\usage{
cities.list(profileId, countryDartIds = NULL, dartIds = NULL,
  namePrefix = NULL, regionDartIds = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{countryDartIds}{Select only cities from these countries}

\item{dartIds}{Select only cities with these DART IDs}

\item{namePrefix}{Select only cities with names starting with this prefix}

\item{regionDartIds}{Select only cities from these regions}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}
