% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{City}
\alias{City}
\title{City Object}
\usage{
City(countryCode = NULL, countryDartId = NULL, dartId = NULL,
  metroCode = NULL, metroDmaId = NULL, name = NULL, regionCode = NULL,
  regionDartId = NULL)
}
\arguments{
\item{countryCode}{Country code of the country to which this city belongs}

\item{countryDartId}{DART ID of the country to which this city belongs}

\item{dartId}{DART ID of this city}

\item{metroCode}{Metro region code of the metro region (DMA) to which this city belongs}

\item{metroDmaId}{ID of the metro region (DMA) to which this city belongs}

\item{name}{Name of this city}

\item{regionCode}{Region code of the region to which this city belongs}

\item{regionDartId}{DART ID of the region to which this city belongs}
}
\value{
City object
}
\description{
City Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about a city that can be targeted by ads.
}
