% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ClickThroughUrl}
\alias{ClickThroughUrl}
\title{ClickThroughUrl Object}
\usage{
ClickThroughUrl(computedClickThroughUrl = NULL,
  customClickThroughUrl = NULL, defaultLandingPage = NULL,
  landingPageId = NULL)
}
\arguments{
\item{computedClickThroughUrl}{Read-only convenience field representing the actual URL that will be used for this click-through}

\item{customClickThroughUrl}{Custom click-through URL}

\item{defaultLandingPage}{Whether the campaign default landing page is used}

\item{landingPageId}{ID of the landing page for the click-through URL}
}
\value{
ClickThroughUrl object
}
\description{
ClickThroughUrl Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Click-through URL
}
