% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CompanionSetting}
\alias{CompanionSetting}
\title{CompanionSetting Object}
\usage{
CompanionSetting(companionsDisabled = NULL, enabledSizes = NULL,
  imageOnly = NULL)
}
\arguments{
\item{companionsDisabled}{Whether companions are disabled for this placement}

\item{enabledSizes}{Whitelist of companion sizes to be served to this placement}

\item{imageOnly}{Whether to serve only static images as companions}
}
\value{
CompanionSetting object
}
\description{
CompanionSetting Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Companion Settings
}
