% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ContentCategory}
\alias{ContentCategory}
\title{ContentCategory Object}
\usage{
ContentCategory(accountId = NULL, id = NULL, name = NULL)
}
\arguments{
\item{accountId}{Account ID of this content category}

\item{id}{ID of this content category}

\item{name}{Name of this content category}
}
\value{
ContentCategory object
}
\description{
ContentCategory Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Organizes placements according to the contents of their associated webpages.
}
\seealso{
Other ContentCategory functions: \code{\link{contentCategories.insert}},
  \code{\link{contentCategories.patch}},
  \code{\link{contentCategories.update}}
}
