% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{conversions.batchinsert}
\alias{conversions.batchinsert}
\title{Inserts conversions.}
\usage{
conversions.batchinsert(ConversionsBatchInsertRequest, profileId)
}
\arguments{
\item{ConversionsBatchInsertRequest}{The \link{ConversionsBatchInsertRequest} object to pass to this method}

\item{profileId}{User profile ID associated with this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/ddmconversions
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/ddmconversions)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}

Other ConversionsBatchInsertRequest functions: \code{\link{ConversionsBatchInsertRequest}}
}
