% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ConversionStatus}
\alias{ConversionStatus}
\title{ConversionStatus Object}
\usage{
ConversionStatus(conversion = NULL, errors = NULL)
}
\arguments{
\item{conversion}{The original conversion that was inserted or updated}

\item{errors}{A list of errors related to this conversion}
}
\value{
ConversionStatus object
}
\description{
ConversionStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The original conversion that was inserted or updated and whether there were any errors.
}
