% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CountriesListResponse}
\alias{CountriesListResponse}
\title{CountriesListResponse Object}
\usage{
CountriesListResponse(countries = NULL)
}
\arguments{
\item{countries}{Country collection}
}
\value{
CountriesListResponse object
}
\description{
CountriesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Country List Response
}
