% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CreativeAssetId}
\alias{CreativeAssetId}
\title{CreativeAssetId Object}
\usage{
CreativeAssetId(name = NULL, type = NULL)
}
\arguments{
\item{name}{Name of the creative asset}

\item{type}{Type of asset to upload}
}
\value{
CreativeAssetId object
}
\description{
CreativeAssetId Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creative Asset ID.
}
