% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CreativeAssetMetadata}
\alias{CreativeAssetMetadata}
\title{CreativeAssetMetadata Object}
\usage{
CreativeAssetMetadata(assetIdentifier = NULL, clickTags = NULL,
  detectedFeatures = NULL, id = NULL, idDimensionValue = NULL,
  warnedValidationRules = NULL)
}
\arguments{
\item{assetIdentifier}{ID of the creative asset}

\item{clickTags}{List of detected click tags for assets}

\item{detectedFeatures}{List of feature dependencies for the creative asset that are detected by DCM}

\item{id}{Numeric ID of the asset}

\item{idDimensionValue}{Dimension value for the numeric ID of the asset}

\item{warnedValidationRules}{Rules validated during code generation that generated a warning}
}
\value{
CreativeAssetMetadata object
}
\description{
CreativeAssetMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
CreativeAssets contains properties of a creative asset file which will be uploaded or has already been uploaded. Refer to the creative sample code for how to upload assets and insert a creative.
}
\seealso{
Other CreativeAssetMetadata functions: \code{\link{creativeAssets.insert}}
}
