% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CreativeAssetSelection}
\alias{CreativeAssetSelection}
\title{CreativeAssetSelection Object}
\usage{
CreativeAssetSelection(defaultAssetId = NULL, rules = NULL)
}
\arguments{
\item{defaultAssetId}{A creativeAssets[]}

\item{rules}{Rules determine which asset will be served to a viewer}
}
\value{
CreativeAssetSelection object
}
\description{
CreativeAssetSelection Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Encapsulates the list of rules for asset selection and a default asset in case none of the rules match. Applicable to INSTREAM_VIDEO creatives.
}
