% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CreativeAssignment}
\alias{CreativeAssignment}
\title{CreativeAssignment Object}
\usage{
CreativeAssignment(active = NULL, applyEventTags = NULL,
  clickThroughUrl = NULL, companionCreativeOverrides = NULL,
  creativeGroupAssignments = NULL, creativeId = NULL,
  creativeIdDimensionValue = NULL, endTime = NULL,
  richMediaExitOverrides = NULL, sequence = NULL, sslCompliant = NULL,
  startTime = NULL, weight = NULL)
}
\arguments{
\item{active}{Whether this creative assignment is active}

\item{applyEventTags}{Whether applicable event tags should fire when this creative assignment is rendered}

\item{clickThroughUrl}{Click-through URL of the creative assignment}

\item{companionCreativeOverrides}{Companion creative overrides for this creative assignment}

\item{creativeGroupAssignments}{Creative group assignments for this creative assignment}

\item{creativeId}{ID of the creative to be assigned}

\item{creativeIdDimensionValue}{Dimension value for the ID of the creative}

\item{endTime}{Date and time that the assigned creative should stop serving}

\item{richMediaExitOverrides}{Rich media exit overrides for this creative assignment}

\item{sequence}{Sequence number of the creative assignment, applicable when the rotation type is CREATIVE_ROTATION_TYPE_SEQUENTIAL}

\item{sslCompliant}{Whether the creative to be assigned is SSL-compliant}

\item{startTime}{Date and time that the assigned creative should start serving}

\item{weight}{Weight of the creative assignment, applicable when the rotation type is CREATIVE_ROTATION_TYPE_RANDOM}
}
\value{
CreativeAssignment object
}
\description{
CreativeAssignment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creative Assignment.
}
