% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CreativeField}
\alias{CreativeField}
\title{CreativeField Object}
\usage{
CreativeField(accountId = NULL, advertiserId = NULL,
  advertiserIdDimensionValue = NULL, id = NULL, name = NULL,
  subaccountId = NULL)
}
\arguments{
\item{accountId}{Account ID of this creative field}

\item{advertiserId}{Advertiser ID of this creative field}

\item{advertiserIdDimensionValue}{Dimension value for the ID of the advertiser}

\item{id}{ID of this creative field}

\item{name}{Name of this creative field}

\item{subaccountId}{Subaccount ID of this creative field}
}
\value{
CreativeField object
}
\description{
CreativeField Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a creative field.
}
\seealso{
Other CreativeField functions: \code{\link{creativeFields.insert}},
  \code{\link{creativeFields.patch}},
  \code{\link{creativeFields.update}}
}
