% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CreativeFieldAssignment}
\alias{CreativeFieldAssignment}
\title{CreativeFieldAssignment Object}
\usage{
CreativeFieldAssignment(creativeFieldId = NULL, creativeFieldValueId = NULL)
}
\arguments{
\item{creativeFieldId}{ID of the creative field}

\item{creativeFieldValueId}{ID of the creative field value}
}
\value{
CreativeFieldAssignment object
}
\description{
CreativeFieldAssignment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creative Field Assignment.
}
