% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CreativeFieldsListResponse}
\alias{CreativeFieldsListResponse}
\title{CreativeFieldsListResponse Object}
\usage{
CreativeFieldsListResponse(creativeFields = NULL, nextPageToken = NULL)
}
\arguments{
\item{creativeFields}{Creative field collection}

\item{nextPageToken}{Pagination token to be used for the next list operation}
}
\value{
CreativeFieldsListResponse object
}
\description{
CreativeFieldsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creative Field List Response
}
