% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CreativeFieldValuesListResponse}
\alias{CreativeFieldValuesListResponse}
\title{CreativeFieldValuesListResponse Object}
\usage{
CreativeFieldValuesListResponse(creativeFieldValues = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{creativeFieldValues}{Creative field value collection}

\item{nextPageToken}{Pagination token to be used for the next list operation}
}
\value{
CreativeFieldValuesListResponse object
}
\description{
CreativeFieldValuesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creative Field Value List Response
}
