% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{creativeGroups.patch}
\alias{creativeGroups.patch}
\title{Updates an existing creative group. This method supports patch semantics.}
\usage{
creativeGroups.patch(CreativeGroup, profileId, id)
}
\arguments{
\item{CreativeGroup}{The \link{CreativeGroup} object to pass to this method}

\item{profileId}{User profile ID associated with this request}

\item{id}{Creative group ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}

Other CreativeGroup functions: \code{\link{CreativeGroup}},
  \code{\link{creativeGroups.insert}},
  \code{\link{creativeGroups.update}}
}
