% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CreativeGroupsListResponse}
\alias{CreativeGroupsListResponse}
\title{CreativeGroupsListResponse Object}
\usage{
CreativeGroupsListResponse(creativeGroups = NULL, nextPageToken = NULL)
}
\arguments{
\item{creativeGroups}{Creative group collection}

\item{nextPageToken}{Pagination token to be used for the next list operation}
}
\value{
CreativeGroupsListResponse object
}
\description{
CreativeGroupsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creative Group List Response
}
