% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CreativeOptimizationConfiguration}
\alias{CreativeOptimizationConfiguration}
\title{CreativeOptimizationConfiguration Object}
\usage{
CreativeOptimizationConfiguration(id = NULL, name = NULL,
  optimizationActivitys = NULL, optimizationModel = NULL)
}
\arguments{
\item{id}{ID of this creative optimization config}

\item{name}{Name of this creative optimization config}

\item{optimizationActivitys}{List of optimization activities associated with this configuration}

\item{optimizationModel}{Optimization model for this configuration}
}
\value{
CreativeOptimizationConfiguration object
}
\description{
CreativeOptimizationConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creative optimization settings.
}
