% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CreativeRotation}
\alias{CreativeRotation}
\title{CreativeRotation Object}
\usage{
CreativeRotation(creativeAssignments = NULL,
  creativeOptimizationConfigurationId = NULL, type = NULL,
  weightCalculationStrategy = NULL)
}
\arguments{
\item{creativeAssignments}{Creative assignments in this creative rotation}

\item{creativeOptimizationConfigurationId}{Creative optimization configuration that is used by this ad}

\item{type}{Type of creative rotation}

\item{weightCalculationStrategy}{Strategy for calculating weights}
}
\value{
CreativeRotation object
}
\description{
CreativeRotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creative Rotation.
}
