% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{creatives.list}
\alias{creatives.list}
\title{Retrieves a list of creatives, possibly filtered. This method supports paging.}
\usage{
creatives.list(profileId, active = NULL, advertiserId = NULL,
  archived = NULL, campaignId = NULL, companionCreativeIds = NULL,
  creativeFieldIds = NULL, ids = NULL, maxResults = NULL,
  pageToken = NULL, renderingIds = NULL, searchString = NULL,
  sizeIds = NULL, sortField = NULL, sortOrder = NULL,
  studioCreativeId = NULL, types = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{active}{Select only active creatives}

\item{advertiserId}{Select only creatives with this advertiser ID}

\item{archived}{Select only archived creatives}

\item{campaignId}{Select only creatives with this campaign ID}

\item{companionCreativeIds}{Select only in-stream video creatives with these companion IDs}

\item{creativeFieldIds}{Select only creatives with these creative field IDs}

\item{ids}{Select only creatives with these IDs}

\item{maxResults}{Maximum number of results to return}

\item{pageToken}{Value of the nextPageToken from the previous result page}

\item{renderingIds}{Select only creatives with these rendering IDs}

\item{searchString}{Allows searching for objects by name or ID}

\item{sizeIds}{Select only creatives with these size IDs}

\item{sortField}{Field by which to sort the list}

\item{sortOrder}{Order of sorted results, default is ASCENDING}

\item{studioCreativeId}{Select only creatives corresponding to this Studio creative ID}

\item{types}{Select only creatives with these creative types}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}
