% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CreativeSettings}
\alias{CreativeSettings}
\title{CreativeSettings Object}
\usage{
CreativeSettings(iFrameFooter = NULL, iFrameHeader = NULL)
}
\arguments{
\item{iFrameFooter}{Header text for iFrames for this site}

\item{iFrameHeader}{Header text for iFrames for this site}
}
\value{
CreativeSettings object
}
\description{
CreativeSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creative Settings
}
