% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CreativesListResponse}
\alias{CreativesListResponse}
\title{CreativesListResponse Object}
\usage{
CreativesListResponse(creatives = NULL, nextPageToken = NULL)
}
\arguments{
\item{creatives}{Creative collection}

\item{nextPageToken}{Pagination token to be used for the next list operation}
}
\value{
CreativesListResponse object
}
\description{
CreativesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creative List Response
}
