% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CrossDimensionReachReportCompatibleFields}
\alias{CrossDimensionReachReportCompatibleFields}
\title{CrossDimensionReachReportCompatibleFields Object}
\usage{
CrossDimensionReachReportCompatibleFields(breakdown = NULL,
  dimensionFilters = NULL, metrics = NULL, overlapMetrics = NULL)
}
\arguments{
\item{breakdown}{Dimensions which are compatible to be selected in the 'breakdown' section of the report}

\item{dimensionFilters}{Dimensions which are compatible to be selected in the 'dimensionFilters' section of the report}

\item{metrics}{Metrics which are compatible to be selected in the 'metricNames' section of the report}

\item{overlapMetrics}{Metrics which are compatible to be selected in the 'overlapMetricNames' section of the report}
}
\value{
CrossDimensionReachReportCompatibleFields object
}
\description{
CrossDimensionReachReportCompatibleFields Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents fields that are compatible to be selected for a report of type 'CROSS_DIMENSION_REACH'.
}
