% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CustomFloodlightVariable}
\alias{CustomFloodlightVariable}
\title{CustomFloodlightVariable Object}
\usage{
CustomFloodlightVariable(type = NULL, value = NULL)
}
\arguments{
\item{type}{The type of custom floodlight variable to supply a value for}

\item{value}{The value of the custom floodlight variable}
}
\value{
CustomFloodlightVariable object
}
\description{
CustomFloodlightVariable Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A custom floodlight variable.
}
