% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{DateRange}
\alias{DateRange}
\title{DateRange Object}
\usage{
DateRange(endDate = NULL, relativeDateRange = NULL, startDate = NULL)
}
\arguments{
\item{endDate}{The end date of the date range, inclusive}

\item{relativeDateRange}{The date range relative to the date of when the report is run}

\item{startDate}{The start date of the date range, inclusive}
}
\value{
DateRange object
}
\description{
DateRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a date range.
}
