% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{DayPartTargeting}
\alias{DayPartTargeting}
\title{DayPartTargeting Object}
\usage{
DayPartTargeting(daysOfWeek = NULL, hoursOfDay = NULL,
  userLocalTime = NULL)
}
\arguments{
\item{daysOfWeek}{Days of the week when the ad will serve}

\item{hoursOfDay}{Hours of the day when the ad will serve}

\item{userLocalTime}{Whether or not to use the user's local time}
}
\value{
DayPartTargeting object
}
\description{
DayPartTargeting Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Day Part Targeting.
}
