% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{DefaultClickThroughEventTagProperties}
\alias{DefaultClickThroughEventTagProperties}
\title{DefaultClickThroughEventTagProperties Object}
\usage{
DefaultClickThroughEventTagProperties(defaultClickThroughEventTagId = NULL,
  overrideInheritedEventTag = NULL)
}
\arguments{
\item{defaultClickThroughEventTagId}{ID of the click-through event tag to apply to all ads in this entity's scope}

\item{overrideInheritedEventTag}{Whether this entity should override the inherited default click-through event tag with its own defined value}
}
\value{
DefaultClickThroughEventTagProperties object
}
\description{
DefaultClickThroughEventTagProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Properties of inheriting and overriding the default click-through event tag. A campaign may override the event tag defined at the advertiser level, and an ad may also override the campaign's setting further.
}
