% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{DeliverySchedule}
\alias{DeliverySchedule}
\title{DeliverySchedule Object}
\usage{
DeliverySchedule(frequencyCap = NULL, hardCutoff = NULL,
  impressionRatio = NULL, priority = NULL)
}
\arguments{
\item{frequencyCap}{Limit on the number of times an individual user can be served the ad within a specified period of time}

\item{hardCutoff}{Whether or not hard cutoff is enabled}

\item{impressionRatio}{Impression ratio for this ad}

\item{priority}{Serving priority of an ad, with respect to other ads}
}
\value{
DeliverySchedule object
}
\description{
DeliverySchedule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Delivery Schedule.
}
