% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Dimension}
\alias{Dimension}
\title{Dimension Object}
\usage{
Dimension(name = NULL)
}
\arguments{
\item{name}{The dimension name, e}
}
\value{
Dimension object
}
\description{
Dimension Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a dimension.
}
