% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{DimensionValue}
\alias{DimensionValue}
\title{DimensionValue Object}
\usage{
DimensionValue(dimensionName = NULL, etag = NULL, id = NULL,
  matchType = NULL, value = NULL)
}
\arguments{
\item{dimensionName}{The name of the dimension}

\item{etag}{The eTag of this response for caching purposes}

\item{id}{The ID associated with the value if available}

\item{matchType}{Determines how the 'value' field is matched when filtering}

\item{value}{The value of the dimension}
}
\value{
DimensionValue object
}
\description{
DimensionValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a DimensionValue resource.
}
