% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{DirectorySite}
\alias{DirectorySite}
\title{DirectorySite Object}
\usage{
DirectorySite(active = NULL, contactAssignments = NULL, countryId = NULL,
  currencyId = NULL, description = NULL, id = NULL,
  idDimensionValue = NULL, inpageTagFormats = NULL,
  interstitialTagFormats = NULL, name = NULL, parentId = NULL,
  settings = NULL, url = NULL)
}
\arguments{
\item{active}{Whether this directory site is active}

\item{contactAssignments}{Directory site contacts}

\item{countryId}{Country ID of this directory site}

\item{currencyId}{Currency ID of this directory site}

\item{description}{Description of this directory site}

\item{id}{ID of this directory site}

\item{idDimensionValue}{Dimension value for the ID of this directory site}

\item{inpageTagFormats}{Tag types for regular placements}

\item{interstitialTagFormats}{Tag types for interstitial placements}

\item{name}{Name of this directory site}

\item{parentId}{Parent directory site ID}

\item{settings}{Directory site settings}

\item{url}{URL of this directory site}
}
\value{
DirectorySite object
}
\description{
DirectorySite Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
DirectorySites contains properties of a website from the Site Directory. Sites need to be added to an account via the Sites resource before they can be assigned to a placement.
}
\seealso{
Other DirectorySite functions: \code{\link{directorySites.insert}}
}
