% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{directorySiteContacts.list}
\alias{directorySiteContacts.list}
\title{Retrieves a list of directory site contacts, possibly filtered. This method supports paging.}
\usage{
directorySiteContacts.list(profileId, directorySiteIds = NULL, ids = NULL,
  maxResults = NULL, pageToken = NULL, searchString = NULL,
  sortField = NULL, sortOrder = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{directorySiteIds}{Select only directory site contacts with these directory site IDs}

\item{ids}{Select only directory site contacts with these IDs}

\item{maxResults}{Maximum number of results to return}

\item{pageToken}{Value of the nextPageToken from the previous result page}

\item{searchString}{Allows searching for objects by name, ID or email}

\item{sortField}{Field by which to sort the list}

\item{sortOrder}{Order of sorted results, default is ASCENDING}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}
