% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{directorySites.list}
\alias{directorySites.list}
\title{Retrieves a list of directory sites, possibly filtered. This method supports paging.}
\usage{
directorySites.list(profileId, acceptsInStreamVideoPlacements = NULL,
  acceptsInterstitialPlacements = NULL,
  acceptsPublisherPaidPlacements = NULL, active = NULL, countryId = NULL,
  dfp_network_code = NULL, ids = NULL, maxResults = NULL,
  pageToken = NULL, parentId = NULL, searchString = NULL,
  sortField = NULL, sortOrder = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{acceptsInStreamVideoPlacements}{This search filter is no longer supported and will have no effect on the results returned}

\item{acceptsInterstitialPlacements}{This search filter is no longer supported and will have no effect on the results returned}

\item{acceptsPublisherPaidPlacements}{Select only directory sites that accept publisher paid placements}

\item{active}{Select only active directory sites}

\item{countryId}{Select only directory sites with this country ID}

\item{dfp_network_code}{Select only directory sites with this DFP network code}

\item{ids}{Select only directory sites with these IDs}

\item{maxResults}{Maximum number of results to return}

\item{pageToken}{Value of the nextPageToken from the previous result page}

\item{parentId}{Select only directory sites with this parent ID}

\item{searchString}{Allows searching for objects by name, ID or URL}

\item{sortField}{Field by which to sort the list}

\item{sortOrder}{Order of sorted results, default is ASCENDING}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}
