% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{dynamicTargetingKeys.delete}
\alias{dynamicTargetingKeys.delete}
\title{Deletes an existing dynamic targeting key.}
\usage{
dynamicTargetingKeys.delete(profileId, objectId, name, objectType)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{objectId}{ID of the object of this dynamic targeting key}

\item{name}{Name of this dynamic targeting key}

\item{objectType}{Type of the object of this dynamic targeting key}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}
