% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{EncryptionInfo}
\alias{EncryptionInfo}
\title{EncryptionInfo Object}
\usage{
EncryptionInfo(encryptionEntityId = NULL, encryptionEntityType = NULL,
  encryptionSource = NULL)
}
\arguments{
\item{encryptionEntityId}{The encryption entity ID}

\item{encryptionEntityType}{The encryption entity type}

\item{encryptionSource}{Describes whether the encrypted cookie was received from ad serving (the %m macro) or from Data Transfer}
}
\value{
EncryptionInfo object
}
\description{
EncryptionInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A description of how user IDs are encrypted.
}
