% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{EventTagOverride}
\alias{EventTagOverride}
\title{EventTagOverride Object}
\usage{
EventTagOverride(enabled = NULL, id = NULL)
}
\arguments{
\item{enabled}{Whether this override is enabled}

\item{id}{ID of this event tag override}
}
\value{
EventTagOverride object
}
\description{
EventTagOverride Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Event tag override information.
}
