% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{files.get}
\alias{files.get}
\title{Retrieves a report file by its report ID and file ID.}
\usage{
files.get(reportId, fileId)
}
\arguments{
\item{reportId}{The ID of the report}

\item{fileId}{The ID of the report file}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfareporting
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfareporting)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}
