% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Flight}
\alias{Flight}
\title{Flight Object}
\usage{
Flight(endDate = NULL, rateOrCost = NULL, startDate = NULL,
  units = NULL)
}
\arguments{
\item{endDate}{Inventory item flight end date}

\item{rateOrCost}{Rate or cost of this flight}

\item{startDate}{Inventory item flight start date}

\item{units}{Units of this flight}
}
\value{
Flight object
}
\description{
Flight Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Flight
}
