% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{FloodlightActivitiesGenerateTagResponse}
\alias{FloodlightActivitiesGenerateTagResponse}
\title{FloodlightActivitiesGenerateTagResponse Object}
\usage{
FloodlightActivitiesGenerateTagResponse(floodlightActivityTag = NULL)
}
\arguments{
\item{floodlightActivityTag}{Generated tag for this floodlight activity}
}
\value{
FloodlightActivitiesGenerateTagResponse object
}
\description{
FloodlightActivitiesGenerateTagResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Floodlight Activity GenerateTag Response
}
