% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{floodlightActivityGroups.list}
\alias{floodlightActivityGroups.list}
\title{Retrieves a list of floodlight activity groups, possibly filtered. This method supports paging.}
\usage{
floodlightActivityGroups.list(profileId, advertiserId = NULL,
  floodlightConfigurationId = NULL, ids = NULL, maxResults = NULL,
  pageToken = NULL, searchString = NULL, sortField = NULL,
  sortOrder = NULL, type = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{advertiserId}{Select only floodlight activity groups with the specified advertiser ID}

\item{floodlightConfigurationId}{Select only floodlight activity groups with the specified floodlight configuration ID}

\item{ids}{Select only floodlight activity groups with the specified IDs}

\item{maxResults}{Maximum number of results to return}

\item{pageToken}{Value of the nextPageToken from the previous result page}

\item{searchString}{Allows searching for objects by name or ID}

\item{sortField}{Field by which to sort the list}

\item{sortOrder}{Order of sorted results, default is ASCENDING}

\item{type}{Select only floodlight activity groups with the specified floodlight activity group type}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}
