% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{floodlightConfigurations.patch}
\alias{floodlightConfigurations.patch}
\title{Updates an existing floodlight configuration. This method supports patch semantics.}
\usage{
floodlightConfigurations.patch(FloodlightConfiguration, profileId, id)
}
\arguments{
\item{FloodlightConfiguration}{The \link{FloodlightConfiguration} object to pass to this method}

\item{profileId}{User profile ID associated with this request}

\item{id}{Floodlight configuration ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}

Other FloodlightConfiguration functions: \code{\link{FloodlightConfiguration}},
  \code{\link{floodlightConfigurations.update}}
}
