% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{FloodlightReportCompatibleFields}
\alias{FloodlightReportCompatibleFields}
\title{FloodlightReportCompatibleFields Object}
\usage{
FloodlightReportCompatibleFields(dimensionFilters = NULL, dimensions = NULL,
  metrics = NULL)
}
\arguments{
\item{dimensionFilters}{Dimensions which are compatible to be selected in the 'dimensionFilters' section of the report}

\item{dimensions}{Dimensions which are compatible to be selected in the 'dimensions' section of the report}

\item{metrics}{Metrics which are compatible to be selected in the 'metricNames' section of the report}
}
\value{
FloodlightReportCompatibleFields object
}
\description{
FloodlightReportCompatibleFields Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents fields that are compatible to be selected for a report of type 'FlOODLIGHT'.
}
