% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{FsCommand}
\alias{FsCommand}
\title{FsCommand Object}
\usage{
FsCommand(left = NULL, positionOption = NULL, top = NULL,
  windowHeight = NULL, windowWidth = NULL)
}
\arguments{
\item{left}{Distance from the left of the browser}

\item{positionOption}{Position in the browser where the window will open}

\item{top}{Distance from the top of the browser}

\item{windowHeight}{Height of the window}

\item{windowWidth}{Width of the window}
}
\value{
FsCommand object
}
\description{
FsCommand Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
FsCommand.
}
