% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{GeoTargeting}
\alias{GeoTargeting}
\title{GeoTargeting Object}
\usage{
GeoTargeting(cities = NULL, countries = NULL, excludeCountries = NULL,
  metros = NULL, postalCodes = NULL, regions = NULL)
}
\arguments{
\item{cities}{Cities to be targeted}

\item{countries}{Countries to be targeted or excluded from targeting, depending on the setting of the excludeCountries field}

\item{excludeCountries}{Whether or not to exclude the countries in the countries field from targeting}

\item{metros}{Metros to be targeted}

\item{postalCodes}{Postal codes to be targeted}

\item{regions}{Regions to be targeted}
}
\value{
GeoTargeting object
}
\description{
GeoTargeting Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Geographical Targeting.
}
