% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{InventoryItem}
\alias{InventoryItem}
\title{InventoryItem Object}
\usage{
InventoryItem(accountId = NULL, adSlots = NULL, advertiserId = NULL,
  contentCategoryId = NULL, estimatedClickThroughRate = NULL,
  estimatedConversionRate = NULL, id = NULL, inPlan = NULL,
  lastModifiedInfo = NULL, name = NULL, negotiationChannelId = NULL,
  orderId = NULL, placementStrategyId = NULL, pricing = NULL,
  projectId = NULL, rfpId = NULL, siteId = NULL, subaccountId = NULL,
  type = NULL)
}
\arguments{
\item{accountId}{Account ID of this inventory item}

\item{adSlots}{Ad slots of this inventory item}

\item{advertiserId}{Advertiser ID of this inventory item}

\item{contentCategoryId}{Content category ID of this inventory item}

\item{estimatedClickThroughRate}{Estimated click-through rate of this inventory item}

\item{estimatedConversionRate}{Estimated conversion rate of this inventory item}

\item{id}{ID of this inventory item}

\item{inPlan}{Whether this inventory item is in plan}

\item{lastModifiedInfo}{Information about the most recent modification of this inventory item}

\item{name}{Name of this inventory item}

\item{negotiationChannelId}{Negotiation channel ID of this inventory item}

\item{orderId}{Order ID of this inventory item}

\item{placementStrategyId}{Placement strategy ID of this inventory item}

\item{pricing}{Pricing of this inventory item}

\item{projectId}{Project ID of this inventory item}

\item{rfpId}{RFP ID of this inventory item}

\item{siteId}{ID of the site this inventory item is associated with}

\item{subaccountId}{Subaccount ID of this inventory item}

\item{type}{Type of inventory item}
}
\value{
InventoryItem object
}
\description{
InventoryItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a buy from the DoubleClick Planning inventory store.
}
