% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{KeyValueTargetingExpression}
\alias{KeyValueTargetingExpression}
\title{KeyValueTargetingExpression Object}
\usage{
KeyValueTargetingExpression(expression = NULL)
}
\arguments{
\item{expression}{Keyword expression being targeted by the ad}
}
\value{
KeyValueTargetingExpression object
}
\description{
KeyValueTargetingExpression Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Key Value Targeting Expression.
}
