% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{LandingPage}
\alias{LandingPage}
\title{LandingPage Object}
\usage{
LandingPage(default = NULL, id = NULL, name = NULL, url = NULL)
}
\arguments{
\item{default}{Whether or not this landing page will be assigned to any ads or creatives that do not have a landing page assigned explicitly}

\item{id}{ID of this landing page}

\item{name}{Name of this landing page}

\item{url}{URL of this landing page}
}
\value{
LandingPage object
}
\description{
LandingPage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about where a user's browser is taken after the user clicks an ad.
}
\seealso{
Other LandingPage functions: \code{\link{landingPages.insert}},
  \code{\link{landingPages.patch}},
  \code{\link{landingPages.update}}
}
