% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{landingPages.insert}
\alias{landingPages.insert}
\title{Inserts a new landing page for the specified campaign.}
\usage{
landingPages.insert(LandingPage, profileId, campaignId)
}
\arguments{
\item{LandingPage}{The \link{LandingPage} object to pass to this method}

\item{profileId}{User profile ID associated with this request}

\item{campaignId}{Landing page campaign ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}

Other LandingPage functions: \code{\link{LandingPage}},
  \code{\link{landingPages.patch}},
  \code{\link{landingPages.update}}
}
