% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{LanguageTargeting}
\alias{LanguageTargeting}
\title{LanguageTargeting Object}
\usage{
LanguageTargeting(languages = NULL)
}
\arguments{
\item{languages}{Languages that this ad targets}
}
\value{
LanguageTargeting object
}
\description{
LanguageTargeting Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Language Targeting.
}
