% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ListPopulationClause}
\alias{ListPopulationClause}
\title{ListPopulationClause Object}
\usage{
ListPopulationClause(terms = NULL)
}
\arguments{
\item{terms}{Terms of this list population clause}
}
\value{
ListPopulationClause object
}
\description{
ListPopulationClause Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A group clause made up of list population terms representing constraints joined by ORs.
}
