% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ListPopulationRule}
\alias{ListPopulationRule}
\title{ListPopulationRule Object}
\usage{
ListPopulationRule(floodlightActivityId = NULL,
  floodlightActivityName = NULL, listPopulationClauses = NULL)
}
\arguments{
\item{floodlightActivityId}{Floodlight activity ID associated with this rule}

\item{floodlightActivityName}{Name of floodlight activity associated with this rule}

\item{listPopulationClauses}{Clauses that make up this list population rule}
}
\value{
ListPopulationRule object
}
\description{
ListPopulationRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Remarketing List Population Rule.
}
