% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ListPopulationTerm}
\alias{ListPopulationTerm}
\title{ListPopulationTerm Object}
\usage{
ListPopulationTerm(contains = NULL, negation = NULL, operator = NULL,
  remarketingListId = NULL, type = NULL, value = NULL,
  variableFriendlyName = NULL, variableName = NULL)
}
\arguments{
\item{contains}{Will be true if the term should check if the user is in the list and false if the term should check if the user is not in the list}

\item{negation}{Whether to negate the comparison result of this term during rule evaluation}

\item{operator}{Comparison operator of this term}

\item{remarketingListId}{ID of the list in question}

\item{type}{List population term type determines the applicable fields in this object}

\item{value}{Literal to compare the variable to}

\item{variableFriendlyName}{Friendly name of this term's variable}

\item{variableName}{Name of the variable (U1, U2, etc}
}
\value{
ListPopulationTerm object
}
\description{
ListPopulationTerm Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Remarketing List Population Rule Term.
}
