% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{LookbackConfiguration}
\alias{LookbackConfiguration}
\title{LookbackConfiguration Object}
\usage{
LookbackConfiguration(clickDuration = NULL,
  postImpressionActivitiesDuration = NULL)
}
\arguments{
\item{clickDuration}{Lookback window, in days, from the last time a given user clicked on one of your ads}

\item{postImpressionActivitiesDuration}{Lookback window, in days, from the last time a given user viewed one of your ads}
}
\value{
LookbackConfiguration object
}
\description{
LookbackConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Lookback configuration settings.
}
