% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Metric}
\alias{Metric}
\title{Metric Object}
\usage{
Metric(name = NULL)
}
\arguments{
\item{name}{The metric name, e}
}
\value{
Metric object
}
\description{
Metric Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a metric.
}
