% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ObjectFilter}
\alias{ObjectFilter}
\title{ObjectFilter Object}
\usage{
ObjectFilter(objectIds = NULL, status = NULL)
}
\arguments{
\item{objectIds}{Applicable when status is ASSIGNED}

\item{status}{Status of the filter}
}
\value{
ObjectFilter object
}
\description{
ObjectFilter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Object Filter.
}
