% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{OmnitureSettings}
\alias{OmnitureSettings}
\title{OmnitureSettings Object}
\usage{
OmnitureSettings(omnitureCostDataEnabled = NULL,
  omnitureIntegrationEnabled = NULL)
}
\arguments{
\item{omnitureCostDataEnabled}{Whether placement cost data will be sent to Omniture}

\item{omnitureIntegrationEnabled}{Whether Omniture integration is enabled}
}
\value{
OmnitureSettings object
}
\description{
OmnitureSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Omniture Integration Settings.
}
