% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{OperatingSystemVersion}
\alias{OperatingSystemVersion}
\title{OperatingSystemVersion Object}
\usage{
OperatingSystemVersion(id = NULL, majorVersion = NULL,
  minorVersion = NULL, name = NULL, operatingSystem = NULL)
}
\arguments{
\item{id}{ID of this operating system version}

\item{majorVersion}{Major version (leftmost number) of this operating system version}

\item{minorVersion}{Minor version (number after the first dot) of this operating system version}

\item{name}{Name of this operating system version}

\item{operatingSystem}{Operating system of this operating system version}
}
\value{
OperatingSystemVersion object
}
\description{
OperatingSystemVersion Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about a particular version of an operating system that can be targeted by ads.
}
