% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{OrdersListResponse}
\alias{OrdersListResponse}
\title{OrdersListResponse Object}
\usage{
OrdersListResponse(nextPageToken = NULL, orders = NULL)
}
\arguments{
\item{nextPageToken}{Pagination token to be used for the next list operation}

\item{orders}{Order collection}
}
\value{
OrdersListResponse object
}
\description{
OrdersListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Order List Response
}
