% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{PlacementAssignment}
\alias{PlacementAssignment}
\title{PlacementAssignment Object}
\usage{
PlacementAssignment(active = NULL, placementId = NULL,
  placementIdDimensionValue = NULL, sslRequired = NULL)
}
\arguments{
\item{active}{Whether this placement assignment is active}

\item{placementId}{ID of the placement to be assigned}

\item{placementIdDimensionValue}{Dimension value for the ID of the placement}

\item{sslRequired}{Whether the placement to be assigned requires SSL}
}
\value{
PlacementAssignment object
}
\description{
PlacementAssignment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Placement Assignment.
}
