% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{placementGroups.list}
\alias{placementGroups.list}
\title{Retrieves a list of placement groups, possibly filtered. This method supports paging.}
\usage{
placementGroups.list(profileId, advertiserIds = NULL, archived = NULL,
  campaignIds = NULL, contentCategoryIds = NULL, directorySiteIds = NULL,
  ids = NULL, maxEndDate = NULL, maxResults = NULL, maxStartDate = NULL,
  minEndDate = NULL, minStartDate = NULL, pageToken = NULL,
  placementGroupType = NULL, placementStrategyIds = NULL,
  pricingTypes = NULL, searchString = NULL, siteIds = NULL,
  sortField = NULL, sortOrder = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{advertiserIds}{Select only placement groups that belong to these advertisers}

\item{archived}{Select only archived placements}

\item{campaignIds}{Select only placement groups that belong to these campaigns}

\item{contentCategoryIds}{Select only placement groups that are associated with these content categories}

\item{directorySiteIds}{Select only placement groups that are associated with these directory sites}

\item{ids}{Select only placement groups with these IDs}

\item{maxEndDate}{Select only placements or placement groups whose end date is on or before the specified maxEndDate}

\item{maxResults}{Maximum number of results to return}

\item{maxStartDate}{Select only placements or placement groups whose start date is on or before the specified maxStartDate}

\item{minEndDate}{Select only placements or placement groups whose end date is on or after the specified minEndDate}

\item{minStartDate}{Select only placements or placement groups whose start date is on or after the specified minStartDate}

\item{pageToken}{Value of the nextPageToken from the previous result page}

\item{placementGroupType}{Select only placement groups belonging with this group type}

\item{placementStrategyIds}{Select only placement groups that are associated with these placement strategies}

\item{pricingTypes}{Select only placement groups with these pricing types}

\item{searchString}{Allows searching for placement groups by name or ID}

\item{siteIds}{Select only placement groups that are associated with these sites}

\item{sortField}{Field by which to sort the list}

\item{sortOrder}{Order of sorted results, default is ASCENDING}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}
