% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{PlacementGroupsListResponse}
\alias{PlacementGroupsListResponse}
\title{PlacementGroupsListResponse Object}
\usage{
PlacementGroupsListResponse(nextPageToken = NULL, placementGroups = NULL)
}
\arguments{
\item{nextPageToken}{Pagination token to be used for the next list operation}

\item{placementGroups}{Placement group collection}
}
\value{
PlacementGroupsListResponse object
}
\description{
PlacementGroupsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Placement Group List Response
}
