% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{PlacementTag}
\alias{PlacementTag}
\title{PlacementTag Object}
\usage{
PlacementTag(placementId = NULL, tagDatas = NULL)
}
\arguments{
\item{placementId}{Placement ID}

\item{tagDatas}{Tags generated for this placement}
}
\value{
PlacementTag object
}
\description{
PlacementTag Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Placement Tag
}
