% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{PlatformType}
\alias{PlatformType}
\title{PlatformType Object}
\usage{
PlatformType(id = NULL, name = NULL)
}
\arguments{
\item{id}{ID of this platform type}

\item{name}{Name of this platform type}
}
\value{
PlatformType object
}
\description{
PlatformType Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about a platform type that can be targeted by ads.
}
