% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{PlatformTypesListResponse}
\alias{PlatformTypesListResponse}
\title{PlatformTypesListResponse Object}
\usage{
PlatformTypesListResponse(platformTypes = NULL)
}
\arguments{
\item{platformTypes}{Platform type collection}
}
\value{
PlatformTypesListResponse object
}
\description{
PlatformTypesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Platform Type List Response
}
