% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{PopupWindowProperties}
\alias{PopupWindowProperties}
\title{PopupWindowProperties Object}
\usage{
PopupWindowProperties(dimension = NULL, offset = NULL,
  positionType = NULL, showAddressBar = NULL, showMenuBar = NULL,
  showScrollBar = NULL, showStatusBar = NULL, showToolBar = NULL,
  title = NULL)
}
\arguments{
\item{dimension}{Popup dimension for a creative}

\item{offset}{Upper-left corner coordinates of the popup window}

\item{positionType}{Popup window position either centered or at specific coordinate}

\item{showAddressBar}{Whether to display the browser address bar}

\item{showMenuBar}{Whether to display the browser menu bar}

\item{showScrollBar}{Whether to display the browser scroll bar}

\item{showStatusBar}{Whether to display the browser status bar}

\item{showToolBar}{Whether to display the browser tool bar}

\item{title}{Title of popup window}
}
\value{
PopupWindowProperties object
}
\description{
PopupWindowProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Popup Window Properties.
}
