% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{PostalCodesListResponse}
\alias{PostalCodesListResponse}
\title{PostalCodesListResponse Object}
\usage{
PostalCodesListResponse(postalCodes = NULL)
}
\arguments{
\item{postalCodes}{Postal code collection}
}
\value{
PostalCodesListResponse object
}
\description{
PostalCodesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Postal Code List Response
}
