% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ReachReportCompatibleFields}
\alias{ReachReportCompatibleFields}
\title{ReachReportCompatibleFields Object}
\usage{
ReachReportCompatibleFields(dimensionFilters = NULL, dimensions = NULL,
  metrics = NULL, pivotedActivityMetrics = NULL,
  reachByFrequencyMetrics = NULL)
}
\arguments{
\item{dimensionFilters}{Dimensions which are compatible to be selected in the 'dimensionFilters' section of the report}

\item{dimensions}{Dimensions which are compatible to be selected in the 'dimensions' section of the report}

\item{metrics}{Metrics which are compatible to be selected in the 'metricNames' section of the report}

\item{pivotedActivityMetrics}{Metrics which are compatible to be selected as activity metrics to pivot on in the 'activities' section of the report}

\item{reachByFrequencyMetrics}{Metrics which are compatible to be selected in the 'reachByFrequencyMetricNames' section of the report}
}
\value{
ReachReportCompatibleFields object
}
\description{
ReachReportCompatibleFields Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents fields that are compatible to be selected for a report of type 'REACH'.
}
