% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Region}
\alias{Region}
\title{Region Object}
\usage{
Region(countryCode = NULL, countryDartId = NULL, dartId = NULL,
  name = NULL, regionCode = NULL)
}
\arguments{
\item{countryCode}{Country code of the country to which this region belongs}

\item{countryDartId}{DART ID of the country to which this region belongs}

\item{dartId}{DART ID of this region}

\item{name}{Name of this region}

\item{regionCode}{Region code}
}
\value{
Region object
}
\description{
Region Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about a region that can be targeted by ads.
}
