% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{RegionsListResponse}
\alias{RegionsListResponse}
\title{RegionsListResponse Object}
\usage{
RegionsListResponse(regions = NULL)
}
\arguments{
\item{regions}{Region collection}
}
\value{
RegionsListResponse object
}
\description{
RegionsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Region List Response
}
