% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{RemarketingList}
\alias{RemarketingList}
\title{RemarketingList Object}
\usage{
RemarketingList(accountId = NULL, active = NULL, advertiserId = NULL,
  advertiserIdDimensionValue = NULL, description = NULL, id = NULL,
  lifeSpan = NULL, listPopulationRule = NULL, listSize = NULL,
  listSource = NULL, name = NULL, subaccountId = NULL)
}
\arguments{
\item{accountId}{Account ID of this remarketing list}

\item{active}{Whether this remarketing list is active}

\item{advertiserId}{Dimension value for the advertiser ID that owns this remarketing list}

\item{advertiserIdDimensionValue}{Dimension value for the ID of the advertiser}

\item{description}{Remarketing list description}

\item{id}{Remarketing list ID}

\item{lifeSpan}{Number of days that a user should remain in the remarketing list without an impression}

\item{listPopulationRule}{Rule used to populate the remarketing list with users}

\item{listSize}{Number of users currently in the list}

\item{listSource}{Product from which this remarketing list was originated}

\item{name}{Name of the remarketing list}

\item{subaccountId}{Subaccount ID of this remarketing list}
}
\value{
RemarketingList object
}
\description{
RemarketingList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a remarketing list. Remarketing enables you to create lists of users who have performed specific actions on a site, then target ads to members of those lists. This resource can be used to manage remarketing lists that are owned by your advertisers. To see all remarketing lists that are visible to your advertisers, including those that are shared to your advertiser or account, use the TargetableRemarketingLists resource.
}
\seealso{
Other RemarketingList functions: \code{\link{remarketingLists.insert}},
  \code{\link{remarketingLists.patch}},
  \code{\link{remarketingLists.update}}
}
