% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{RemarketingListsListResponse}
\alias{RemarketingListsListResponse}
\title{RemarketingListsListResponse Object}
\usage{
RemarketingListsListResponse(nextPageToken = NULL, remarketingLists = NULL)
}
\arguments{
\item{nextPageToken}{Pagination token to be used for the next list operation}

\item{remarketingLists}{Remarketing list collection}
}
\value{
RemarketingListsListResponse object
}
\description{
RemarketingListsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Remarketing list response
}
