% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Report.criteria}
\alias{Report.criteria}
\title{Report.criteria Object}
\usage{
Report.criteria(activities = NULL, customRichMediaEvents = NULL,
  dateRange = NULL, dimensionFilters = NULL, dimensions = NULL,
  metricNames = NULL)
}
\arguments{
\item{activities}{Activity group}

\item{customRichMediaEvents}{Custom Rich Media Events group}

\item{dateRange}{The date range for which this report should be run}

\item{dimensionFilters}{The list of filters on which dimensions are filtered}

\item{dimensions}{The list of standard dimensions the report should include}

\item{metricNames}{The list of names of metrics the report should include}
}
\value{
Report.criteria object
}
\description{
Report.criteria Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The report criteria for a report of type 'STANDARD'.
}
\seealso{
Other Report functions: \code{\link{Report.crossDimensionReachCriteria}},
  \code{\link{Report.delivery}},
  \code{\link{Report.floodlightCriteria.reportProperties}},
  \code{\link{Report.floodlightCriteria}},
  \code{\link{Report.pathToConversionCriteria.reportProperties}},
  \code{\link{Report.pathToConversionCriteria}},
  \code{\link{Report.reachCriteria}},
  \code{\link{Report.schedule}}, \code{\link{Report}},
  \code{\link{reports.insert}},
  \code{\link{reports.patch}}, \code{\link{reports.update}}
}
