% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Report.crossDimensionReachCriteria}
\alias{Report.crossDimensionReachCriteria}
\title{Report.crossDimensionReachCriteria Object}
\usage{
Report.crossDimensionReachCriteria(breakdown = NULL, dateRange = NULL,
  dimension = NULL, dimensionFilters = NULL, metricNames = NULL,
  overlapMetricNames = NULL, pivoted = NULL)
}
\arguments{
\item{breakdown}{The list of dimensions the report should include}

\item{dateRange}{The date range this report should be run for}

\item{dimension}{The dimension option}

\item{dimensionFilters}{The list of filters on which dimensions are filtered}

\item{metricNames}{The list of names of metrics the report should include}

\item{overlapMetricNames}{The list of names of overlap metrics the report should include}

\item{pivoted}{Whether the report is pivoted or not}
}
\value{
Report.crossDimensionReachCriteria object
}
\description{
Report.crossDimensionReachCriteria Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The report criteria for a report of type 'CROSS_DIMENSION_REACH'.
}
\seealso{
Other Report functions: \code{\link{Report.criteria}},
  \code{\link{Report.delivery}},
  \code{\link{Report.floodlightCriteria.reportProperties}},
  \code{\link{Report.floodlightCriteria}},
  \code{\link{Report.pathToConversionCriteria.reportProperties}},
  \code{\link{Report.pathToConversionCriteria}},
  \code{\link{Report.reachCriteria}},
  \code{\link{Report.schedule}}, \code{\link{Report}},
  \code{\link{reports.insert}},
  \code{\link{reports.patch}}, \code{\link{reports.update}}
}
