% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Report.delivery}
\alias{Report.delivery}
\title{Report.delivery Object}
\usage{
Report.delivery(emailOwner = NULL, emailOwnerDeliveryType = NULL,
  message = NULL, recipients = NULL)
}
\arguments{
\item{emailOwner}{Whether the report should be emailed to the report owner}

\item{emailOwnerDeliveryType}{The type of delivery for the owner to receive, if enabled}

\item{message}{The message to be sent with each email}

\item{recipients}{The list of recipients to which to email the report}
}
\value{
Report.delivery object
}
\description{
Report.delivery Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The report's email delivery settings.
}
\seealso{
Other Report functions: \code{\link{Report.criteria}},
  \code{\link{Report.crossDimensionReachCriteria}},
  \code{\link{Report.floodlightCriteria.reportProperties}},
  \code{\link{Report.floodlightCriteria}},
  \code{\link{Report.pathToConversionCriteria.reportProperties}},
  \code{\link{Report.pathToConversionCriteria}},
  \code{\link{Report.reachCriteria}},
  \code{\link{Report.schedule}}, \code{\link{Report}},
  \code{\link{reports.insert}},
  \code{\link{reports.patch}}, \code{\link{reports.update}}
}
