% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Report.floodlightCriteria.reportProperties}
\alias{Report.floodlightCriteria.reportProperties}
\title{Report.floodlightCriteria.reportProperties Object}
\usage{

  Report.floodlightCriteria.reportProperties(includeAttributedIPConversions = NULL,
  includeUnattributedCookieConversions = NULL,
  includeUnattributedIPConversions = NULL)
}
\arguments{
\item{includeAttributedIPConversions}{Include conversions that have no cookie, but do have an exposure path}

\item{includeUnattributedCookieConversions}{Include conversions of users with a DoubleClick cookie but without an exposure}

\item{includeUnattributedIPConversions}{Include conversions that have no associated cookies and no exposures}
}
\value{
Report.floodlightCriteria.reportProperties object
}
\description{
Report.floodlightCriteria.reportProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The properties of the report.
}
\seealso{
Other Report functions: \code{\link{Report.criteria}},
  \code{\link{Report.crossDimensionReachCriteria}},
  \code{\link{Report.delivery}},
  \code{\link{Report.floodlightCriteria}},
  \code{\link{Report.pathToConversionCriteria.reportProperties}},
  \code{\link{Report.pathToConversionCriteria}},
  \code{\link{Report.reachCriteria}},
  \code{\link{Report.schedule}}, \code{\link{Report}},
  \code{\link{reports.insert}},
  \code{\link{reports.patch}}, \code{\link{reports.update}}
}
