% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Report.pathToConversionCriteria.reportProperties}
\alias{Report.pathToConversionCriteria.reportProperties}
\title{Report.pathToConversionCriteria.reportProperties Object}
\usage{
Report.pathToConversionCriteria.reportProperties(clicksLookbackWindow = NULL,
  impressionsLookbackWindow = NULL, includeAttributedIPConversions = NULL,
  includeUnattributedCookieConversions = NULL,
  includeUnattributedIPConversions = NULL, maximumClickInteractions = NULL,
  maximumImpressionInteractions = NULL, maximumInteractionGap = NULL,
  pivotOnInteractionPath = NULL)
}
\arguments{
\item{clicksLookbackWindow}{DFA checks to see if a click interaction occurred within the specified period of time before a conversion}

\item{impressionsLookbackWindow}{DFA checks to see if an impression interaction occurred within the specified period of time before a conversion}

\item{includeAttributedIPConversions}{Deprecated: has no effect}

\item{includeUnattributedCookieConversions}{Include conversions of users with a DoubleClick cookie but without an exposure}

\item{includeUnattributedIPConversions}{Include conversions that have no associated cookies and no exposures}

\item{maximumClickInteractions}{The maximum number of click interactions to include in the report}

\item{maximumImpressionInteractions}{The maximum number of click interactions to include in the report}

\item{maximumInteractionGap}{The maximum amount of time that can take place between interactions (clicks or impressions) by the same user}

\item{pivotOnInteractionPath}{Enable pivoting on interaction path}
}
\value{
Report.pathToConversionCriteria.reportProperties object
}
\description{
Report.pathToConversionCriteria.reportProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The properties of the report.
}
\seealso{
Other Report functions: \code{\link{Report.criteria}},
  \code{\link{Report.crossDimensionReachCriteria}},
  \code{\link{Report.delivery}},
  \code{\link{Report.floodlightCriteria.reportProperties}},
  \code{\link{Report.floodlightCriteria}},
  \code{\link{Report.pathToConversionCriteria}},
  \code{\link{Report.reachCriteria}},
  \code{\link{Report.schedule}}, \code{\link{Report}},
  \code{\link{reports.insert}},
  \code{\link{reports.patch}}, \code{\link{reports.update}}
}
