% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ReportList}
\alias{ReportList}
\title{ReportList Object}
\usage{
ReportList(etag = NULL, items = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{The eTag of this response for caching purposes}

\item{items}{The reports returned in this response}

\item{nextPageToken}{Continuation token used to page through reports}
}
\value{
ReportList object
}
\description{
ReportList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the list of reports.
}
