% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ReportsConfiguration}
\alias{ReportsConfiguration}
\title{ReportsConfiguration Object}
\usage{
ReportsConfiguration(exposureToConversionEnabled = NULL,
  lookbackConfiguration = NULL, reportGenerationTimeZoneId = NULL)
}
\arguments{
\item{exposureToConversionEnabled}{Whether the exposure to conversion report is enabled}

\item{lookbackConfiguration}{Default lookback windows for new advertisers in this account}

\item{reportGenerationTimeZoneId}{Report generation time zone ID of this account}
}
\value{
ReportsConfiguration object
}
\description{
ReportsConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Reporting Configuration
}
