% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Site}
\alias{Site}
\title{Site Object}
\usage{
Site(accountId = NULL, approved = NULL, directorySiteId = NULL,
  directorySiteIdDimensionValue = NULL, id = NULL,
  idDimensionValue = NULL, keyName = NULL, name = NULL,
  siteContacts = NULL, siteSettings = NULL, subaccountId = NULL)
}
\arguments{
\item{accountId}{Account ID of this site}

\item{approved}{Whether this site is approved}

\item{directorySiteId}{Directory site associated with this site}

\item{directorySiteIdDimensionValue}{Dimension value for the ID of the directory site}

\item{id}{ID of this site}

\item{idDimensionValue}{Dimension value for the ID of this site}

\item{keyName}{Key name of this site}

\item{name}{Name of this site}

\item{siteContacts}{Site contacts}

\item{siteSettings}{Site-wide settings}

\item{subaccountId}{Subaccount ID of this site}
}
\value{
Site object
}
\description{
Site Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a site.
}
\seealso{
Other Site functions: \code{\link{sites.insert}},
  \code{\link{sites.patch}}, \code{\link{sites.update}}
}
