% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{SitesListResponse}
\alias{SitesListResponse}
\title{SitesListResponse Object}
\usage{
SitesListResponse(nextPageToken = NULL, sites = NULL)
}
\arguments{
\item{nextPageToken}{Pagination token to be used for the next list operation}

\item{sites}{Site collection}
}
\value{
SitesListResponse object
}
\description{
SitesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Site List Response
}
