% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Size}
\alias{Size}
\title{Size Object}
\usage{
Size(height = NULL, iab = NULL, id = NULL, width = NULL)
}
\arguments{
\item{height}{Height of this size}

\item{iab}{IAB standard size}

\item{id}{ID of this size}

\item{width}{Width of this size}
}
\value{
Size object
}
\description{
Size Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the dimensions of ads, placements, creatives, or creative assets.
}
\seealso{
Other Size functions: \code{\link{sizes.insert}}
}
