% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{SkippableSetting}
\alias{SkippableSetting}
\title{SkippableSetting Object}
\usage{
SkippableSetting(progressOffset = NULL, skipOffset = NULL,
  skippable = NULL)
}
\arguments{
\item{progressOffset}{Amount of time to play videos served to this placement before counting a view}

\item{skipOffset}{Amount of time to play videos served to this placement before the skip button should appear}

\item{skippable}{Whether the user can skip creatives served to this placement}
}
\value{
SkippableSetting object
}
\description{
SkippableSetting Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Skippable Settings
}
