% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{SortedDimension}
\alias{SortedDimension}
\title{SortedDimension Object}
\usage{
SortedDimension(name = NULL, sortOrder = NULL)
}
\arguments{
\item{name}{The name of the dimension}

\item{sortOrder}{An optional sort order for the dimension column}
}
\value{
SortedDimension object
}
\description{
SortedDimension Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a sorted dimension.
}
