% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{SubaccountsListResponse}
\alias{SubaccountsListResponse}
\title{SubaccountsListResponse Object}
\usage{
SubaccountsListResponse(nextPageToken = NULL, subaccounts = NULL)
}
\arguments{
\item{nextPageToken}{Pagination token to be used for the next list operation}

\item{subaccounts}{Subaccount collection}
}
\value{
SubaccountsListResponse object
}
\description{
SubaccountsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Subaccount List Response
}
