% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{TagData}
\alias{TagData}
\title{TagData Object}
\usage{
TagData(adId = NULL, clickTag = NULL, creativeId = NULL, format = NULL,
  impressionTag = NULL)
}
\arguments{
\item{adId}{Ad associated with this placement tag}

\item{clickTag}{Tag string to record a click}

\item{creativeId}{Creative associated with this placement tag}

\item{format}{TagData tag format of this tag}

\item{impressionTag}{Tag string for serving an ad}
}
\value{
TagData object
}
\description{
TagData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Placement Tag Data
}
