% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{TagSettings}
\alias{TagSettings}
\title{TagSettings Object}
\usage{
TagSettings(dynamicTagEnabled = NULL, imageTagEnabled = NULL)
}
\arguments{
\item{dynamicTagEnabled}{Whether dynamic floodlight tags are enabled}

\item{imageTagEnabled}{Whether image tags are enabled}
}
\value{
TagSettings object
}
\description{
TagSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Dynamic and Image Tag Settings.
}
