% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{TargetableRemarketingList}
\alias{TargetableRemarketingList}
\title{TargetableRemarketingList Object}
\usage{
TargetableRemarketingList(accountId = NULL, active = NULL,
  advertiserId = NULL, advertiserIdDimensionValue = NULL,
  description = NULL, id = NULL, lifeSpan = NULL, listSize = NULL,
  listSource = NULL, name = NULL, subaccountId = NULL)
}
\arguments{
\item{accountId}{Account ID of this remarketing list}

\item{active}{Whether this targetable remarketing list is active}

\item{advertiserId}{Dimension value for the advertiser ID that owns this targetable remarketing list}

\item{advertiserIdDimensionValue}{Dimension value for the ID of the advertiser}

\item{description}{Targetable remarketing list description}

\item{id}{Targetable remarketing list ID}

\item{lifeSpan}{Number of days that a user should remain in the targetable remarketing list without an impression}

\item{listSize}{Number of users currently in the list}

\item{listSource}{Product from which this targetable remarketing list was originated}

\item{name}{Name of the targetable remarketing list}

\item{subaccountId}{Subaccount ID of this remarketing list}
}
\value{
TargetableRemarketingList object
}
\description{
TargetableRemarketingList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a targetable remarketing list. Remarketing enables you to create lists of users who have performed specific actions on a site, then target ads to members of those lists. This resource is a read-only view of a remarketing list to be used to faciliate targeting ads to specific lists. Remarketing lists that are owned by your advertisers and those that are shared to your advertisers or account are accessible via this resource. To manage remarketing lists that are owned by your advertisers, use the RemarketingLists resource.
}
