% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{TargetableRemarketingListsListResponse}
\alias{TargetableRemarketingListsListResponse}
\title{TargetableRemarketingListsListResponse Object}
\usage{
TargetableRemarketingListsListResponse(nextPageToken = NULL,
  targetableRemarketingLists = NULL)
}
\arguments{
\item{nextPageToken}{Pagination token to be used for the next list operation}

\item{targetableRemarketingLists}{Targetable remarketing list collection}
}
\value{
TargetableRemarketingListsListResponse object
}
\description{
TargetableRemarketingListsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Targetable remarketing list response
}
