% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{TranscodeSetting}
\alias{TranscodeSetting}
\title{TranscodeSetting Object}
\usage{
TranscodeSetting(enabledVideoFormats = NULL)
}
\arguments{
\item{enabledVideoFormats}{Whitelist of video formats to be served to this placement}
}
\value{
TranscodeSetting object
}
\description{
TranscodeSetting Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Transcode Settings
}
