% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{UserDefinedVariableConfiguration}
\alias{UserDefinedVariableConfiguration}
\title{UserDefinedVariableConfiguration Object}
\usage{
UserDefinedVariableConfiguration(dataType = NULL, reportName = NULL,
  variableType = NULL)
}
\arguments{
\item{dataType}{Data type for the variable}

\item{reportName}{User-friendly name for the variable which will appear in reports}

\item{variableType}{Variable name in the tag}
}
\value{
UserDefinedVariableConfiguration object
}
\description{
UserDefinedVariableConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
User Defined Variable configuration.
}
