% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{UserProfile}
\alias{UserProfile}
\title{UserProfile Object}
\usage{
UserProfile(accountId = NULL, accountName = NULL, etag = NULL,
  profileId = NULL, subAccountId = NULL, subAccountName = NULL,
  userName = NULL)
}
\arguments{
\item{accountId}{The account ID to which this profile belongs}

\item{accountName}{The account name this profile belongs to}

\item{etag}{The eTag of this response for caching purposes}

\item{profileId}{The unique ID of the user profile}

\item{subAccountId}{The sub account ID this profile belongs to if applicable}

\item{subAccountName}{The sub account name this profile belongs to if applicable}

\item{userName}{The user name}
}
\value{
UserProfile object
}
\description{
UserProfile Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a UserProfile resource.
}
