% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{UserRolePermissionGroupsListResponse}
\alias{UserRolePermissionGroupsListResponse}
\title{UserRolePermissionGroupsListResponse Object}
\usage{
UserRolePermissionGroupsListResponse(userRolePermissionGroups = NULL)
}
\arguments{
\item{userRolePermissionGroups}{User role permission group collection}
}
\value{
UserRolePermissionGroupsListResponse object
}
\description{
UserRolePermissionGroupsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
User Role Permission Group List Response
}
