% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{userRolePermissions.list}
\alias{userRolePermissions.list}
\title{Gets a list of user role permissions, possibly filtered.}
\usage{
userRolePermissions.list(profileId, ids = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{ids}{Select only user role permissions with these IDs}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}
