% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{VideoFormat}
\alias{VideoFormat}
\title{VideoFormat Object}
\usage{
VideoFormat(fileType = NULL, id = NULL, resolution = NULL,
  targetBitRate = NULL)
}
\arguments{
\item{fileType}{File type of the video format}

\item{id}{ID of the video format}

\item{resolution}{The resolution of this video format}

\item{targetBitRate}{The target bit rate of this video format}
}
\value{
VideoFormat object
}
\description{
VideoFormat Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about supported video formats.
}
