% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{VideoFormatsListResponse}
\alias{VideoFormatsListResponse}
\title{VideoFormatsListResponse Object}
\usage{
VideoFormatsListResponse(videoFormats = NULL)
}
\arguments{
\item{videoFormats}{Video format collection}
}
\value{
VideoFormatsListResponse object
}
\description{
VideoFormatsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Video Format List Response
}
