% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{VideoOffset}
\alias{VideoOffset}
\title{VideoOffset Object}
\usage{
VideoOffset(offsetPercentage = NULL, offsetSeconds = NULL)
}
\arguments{
\item{offsetPercentage}{Duration, as a percentage of video duration}

\item{offsetSeconds}{Duration, in seconds}
}
\value{
VideoOffset object
}
\description{
VideoOffset Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Video Offset
}
