% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{VideoSettings}
\alias{VideoSettings}
\title{VideoSettings Object}
\usage{
VideoSettings(companionSettings = NULL, skippableSettings = NULL,
  transcodeSettings = NULL)
}
\arguments{
\item{companionSettings}{Settings for the companion creatives of video creatives served to this placement}

\item{skippableSettings}{Settings for the skippability of video creatives served to this placement}

\item{transcodeSettings}{Settings for the transcodes of video creatives served to this placement}
}
\value{
VideoSettings object
}
\description{
VideoSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Video Settings
}
